Moved the TPC part of PHG4SvtxDigitizer from g4simulations/g4hough to g4simulations/tpc. Moved SvtxHit classes to g4detectors, to avoid circular library dependence between g4hough and g4tpc. Cleaned up unused code in the MVTX and INTT digitizers.
