+++ simulation/g4simulation/g4eval/SvtxHitEval.C
@@ -1,7 +1,7 @@
 #include "SvtxHitEval.h"
 
-#include <g4hough/SvtxHitMap.h>
-#include <g4hough/SvtxHit.h>
+#include <g4detectors/SvtxHitMap.h>
+#include <g4detectors/SvtxHit.h>
 #include <g4detectors/PHG4CellContainer.h>
 #include <g4detectors/PHG4Cell.h>
 #include <g4main/PHG4HitContainer.h>
@@ -446,11 +446,10 @@ void SvtxHitEval::get_node_pointers(PHCompositeNode* topNode) {
   _hitmap = findNode::getClass<SvtxHitMap>(topNode,"SvtxHitMap");
 
   // need things off of the DST...
-  _g4cells_svtx    = findNode::getClass<PHG4CellContainer>(topNode,"G4CELL_SVTX");
+  _g4cells_svtx    = findNode::getClass<PHG4CellContainer>(topNode,"G4CELL_TPC");
   _g4cells_tracker = findNode::getClass<PHG4CellContainer>(topNode,"G4CELL_SILICON_TRACKER");
   _g4cells_maps = findNode::getClass<PHG4CellContainer>(topNode,"G4CELL_MAPS");
 
-  //  _g4hits_svtx    = findNode::getClass<PHG4HitContainer>(topNode,"G4HIT_SVTX");
   _g4hits_svtx    = findNode::getClass<PHG4HitContainer>(topNode,"G4HIT_TPC");
   _g4hits_tracker = findNode::getClass<PHG4HitContainer>(topNode,"G4HIT_SILICON_TRACKER");
   _g4hits_maps = findNode::getClass<PHG4HitContainer>(topNode,"G4HIT_MAPS");
